refactor: role constants moved to utils
